code/requirements.txt
altair==4.2.2
matplotlib==3.5.3
nltk==3.8.1
numpy==1.23.5
pandas==1.5.2
Pattern==3.6
Pillow==9.4.0
streamlit==1.20.0
textblob==0.17.1
wordcloud==1.8.2.2
